pyproject.toml
[tool.poetry]
name = "internxt-data-tools"
version = "0.0.1"
description = ""
authors = ["Joan Mora <elcorreudejoanmora@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
stripe = "^2.68.0"
google-cloud-bigquery = "^2.34.2"
pandas = "^1.4.2"
pyarrow = "^7.0.0"
pymongo = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
